clean up, mostly moving constants to helper file
