+++ test/integration/nyql/nyql-precision.test.js
@@ -1,31 +1,5 @@
-const request = require('supertest')
 const { expect } = require('chai')
-const { baseUrl } = require('./helpers')
-
-const normalizeCallnumber = (value) =>
-  String(value || '')
-    .replace(/^[*"]+|[*"]+$/g, '')
-    .trim()
-    .toLowerCase()
-
-const resultContainsCallnumber = (result, callnumber) => {
-  const target = normalizeCallnumber(callnumber)
-  const shelfMarks = [
-    ...(Array.isArray(result?.shelfMark) ? result.shelfMark : []),
-    ...(Array.isArray(result?.items)
-      ? result.items.flatMap((item) =>
-        Array.isArray(item?.shelfMark) ? item.shelfMark : []
-      )
-      : [])
-  ]
-    .filter(Boolean)
-    .map(normalizeCallnumber)
-
-  // Check if any shelf mark contains the target as a substring or exact match
-  return shelfMarks.some(
-    (mark) => mark.includes(target) || target.includes(mark)
-  )
-}
+const { resultContainsCallnumber, search } = require('./helpers')
 
 // These are some tests to verify that certain NYQL queries are returning results with the expected precision, especially for fields like call numbers where we want to ensure that the query is matching the intended values and not over- or under-matching.
 
@@ -33,19 +7,12 @@ describe('Discovery API - NYQL precision tests', function () {
   this.timeout(30000)
 
   it('should return exactly one result for callnumber = "JFE 86-3252"', async () => {
-    const endpoint = '/discovery/resources'
     const callnumber = 'JFE 86-3252'
 
-    const res = await request(baseUrl)
-      .get(endpoint)
-      .query({
-        q: `callnumber = "${callnumber}"`,
-        search_scope: 'cql',
-        per_page: 100
-      })
-      .timeout(30000)
-      .expect(200)
-
+    const res = await search({
+      q: `callnumber = "${callnumber}"`
+    })
+  
     expect(res.body.itemListElement).to.be.an('array')
 
     // Assert exactly one result
@@ -57,18 +24,11 @@ describe('Discovery API - NYQL precision tests', function () {
   })
 
   it('all returned bibs should contain callnumber = "MGZMD"', async () => {
-    const endpoint = '/discovery/resources'
     const callnumber = 'MGZMD'
 
-    const res = await request(baseUrl)
-      .get(endpoint)
-      .query({
-        q: `callnumber = "${callnumber}"`,
-        search_scope: 'cql',
-        per_page: 100
-      })
-      .timeout(30000)
-      .expect(200)
+    const res = await search({
+      q: `callnumber = "${callnumber}"`
+    })
 
     expect(res.body.itemListElement).to.be.an('array')
 
@@ -84,18 +44,11 @@ describe('Discovery API - NYQL precision tests', function () {
   })
 
   it('should return exactly one result for identifier = "b10670401"', async () => {
-    const endpoint = '/discovery/resources'
     const identifier = 'b10670401'
 
-    const res = await request(baseUrl)
-      .get(endpoint)
-      .query({
-        q: `identifier = "${identifier}"`,
-        search_scope: 'cql',
-        per_page: 100
-      })
-      .timeout(30000)
-      .expect(200)
+    const res = await search({
+      q: `identifier = "${identifier}"`
+    })
 
     expect(res.body.itemListElement).to.be.an('array')
